move demand estimation
