--- src/demand_estimation/poverty_levels.csv
@@ -1,10 +1,0 @@
-household_size,one_adult,two_adults
-1,13465.0
-2,17331,17839
-3,20832,20852
-4,26246,26338
-5,30887,30414
-6,34582,33935
-7,38734,37210
-8,42585,42224
-9,57319,57597
